Add v1 of the API
