+++ api/api.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"encoding/json"
 	"github.com/gorilla/mux"
 	"github.com/mp-hl-2021/splinter/api/v1"
 	"github.com/mp-hl-2021/splinter/usecases"
@@ -22,17 +21,4 @@ func (a *Api) Router() http.Handler {
 	a.v1.Router(router.PathPrefix("/api/v1").Subrouter())
 
 	return router
-}
-
-type errorResponse struct {
-	StatusCode int
-	Error      string
-}
-
-func WriteError(w http.ResponseWriter, err error, statusCode int) {
-	w.WriteHeader(statusCode)
-	_ = json.NewEncoder(w).Encode(errorResponse{
-		StatusCode: statusCode,
-		Error:      err.Error(),
-	})
 }
