+++ api/v1/authenticate.go
@@ -5,7 +5,6 @@ package v1
 
 import (
 	"encoding/json"
-	"github.com/mp-hl-2021/splinter/api"
 	"github.com/mp-hl-2021/splinter/usecases"
 	"net/http"
 )
@@ -22,13 +21,13 @@ type authenticateResponse struct {
 func (a *Api) endpointAuthenticate(w http.ResponseWriter, r *http.Request) {
 	var b authenticateBody
 	if err := json.NewDecoder(r.Body).Decode(&b); err != nil {
-		api.WriteError(w, err, http.StatusBadRequest)
+		WriteError(w, err, http.StatusBadRequest)
 		return
 	}
 
 	token, err := a.useCases.Authenticate(b.Username, b.Password)
 	if err != nil {
-		api.WriteError(w, err, http.StatusForbidden)
+		WriteError(w, err, http.StatusForbidden)
 		return
 	}
