+++ api/v1/create_account.go
@@ -5,7 +5,6 @@ package v1
 
 import (
 	"encoding/json"
-	"github.com/mp-hl-2021/splinter/api"
 	"github.com/mp-hl-2021/splinter/usecases"
 	"net/http"
 )
@@ -22,13 +21,13 @@ type createAccountResponse struct {
 func (a *Api) endpointCreateAccount(w http.ResponseWriter, r *http.Request) {
 	var b createAccountBody
 	if err := json.NewDecoder(r.Body).Decode(&b); err != nil {
-		api.WriteError(w, err, http.StatusBadRequest)
+		WriteError(w, err, http.StatusBadRequest)
 		return
 	}
 
 	user, err := a.useCases.CreateAccount(b.Username, b.Password)
 	if err != nil {
-		api.WriteError(w, err, http.StatusBadRequest)
+		WriteError(w, err, http.StatusBadRequest)
 		return
 	}
