+++ api/v1/get_comments.go
@@ -6,7 +6,6 @@ package v1
 import (
 	"encoding/json"
 	"github.com/gorilla/mux"
-	"github.com/mp-hl-2021/splinter/api"
 	"github.com/mp-hl-2021/splinter/usecases"
 	"net/http"
 )
@@ -21,7 +20,7 @@ func (a *Api) endpointGetComments(w http.ResponseWriter, r *http.Request) {
 
 	comments, err := a.useCases.GetComments(snippetId)
 	if err != nil {
-		api.WriteError(w, err, http.StatusNotFound)
+		WriteError(w, err, http.StatusNotFound)
 		return
 	}
