+++ api/v1/get_snippets_by_language.go
@@ -6,7 +6,6 @@ package v1
 import (
 	"encoding/json"
 	"github.com/gorilla/mux"
-	"github.com/mp-hl-2021/splinter/api"
 	"github.com/mp-hl-2021/splinter/usecases"
 	"net/http"
 )
@@ -21,7 +20,7 @@ func (a *Api) endpointGetSnippetsByLanguage(w http.ResponseWriter, r *http.Reque
 
 	snippets, err := a.useCases.GetSnippetsByLanguage(language)
 	if err != nil {
-		api.WriteError(w, err, http.StatusNotFound)
+		WriteError(w, err, http.StatusNotFound)
 		return
 	}
