+++ api/v1/post_comment.go
@@ -6,7 +6,6 @@ package v1
 import (
 	"encoding/json"
 	"github.com/gorilla/mux"
-	"github.com/mp-hl-2021/splinter/api"
 	"github.com/mp-hl-2021/splinter/usecases"
 	"net/http"
 )
@@ -25,13 +24,13 @@ func (a *Api) endpointPostComment(w http.ResponseWriter, r *http.Request) {
 
 	var b postCommentBody
 	if err := json.NewDecoder(r.Body).Decode(&b); err != nil {
-		api.WriteError(w, err, http.StatusBadRequest)
+		WriteError(w, err, http.StatusBadRequest)
 		return
 	}
 
 	Comment, err := a.useCases.PostComment(b.Contents, snippetId)
 	if err != nil {
-		api.WriteError(w, err, http.StatusForbidden)
+		WriteError(w, err, http.StatusForbidden)
 		return
 	}
