+++ api/v1/vote.go
@@ -5,7 +5,6 @@ package v1
 
 import (
 	"encoding/json"
-	"github.com/mp-hl-2021/splinter/api"
 	"github.com/mp-hl-2021/splinter/usecases"
 	"net/http"
 )
@@ -18,12 +17,12 @@ type voteBody struct {
 func (a *Api) endpointVote(w http.ResponseWriter, r *http.Request) {
 	var b voteBody
 	if err := json.NewDecoder(r.Body).Decode(&b); err != nil {
-		api.WriteError(w, err, http.StatusBadRequest)
+		WriteError(w, err, http.StatusBadRequest)
 		return
 	}
 
 	if err := a.useCases.Vote(b.Snippet, b.Vote); err != nil {
-		api.WriteError(w, err, http.StatusNotFound)
+		WriteError(w, err, http.StatusNotFound)
 		return
 	}
